update doc and controller

+++ api/app/Http/Controllers/DataController.php
@@ -9,18 +9,19 @@ use App\Http\Controllers\Controller;
 
 class DataController extends Controller {
     /**
-    *  @api {get} /api/{app_id}/data 获得app数据
+    *  @api {get} /api/app/{app_id}/data 获得app数据
     *  @apiName show_data
     *  @apiGroup Data
-    *  @apiVersion v1.0.0
+    *  @apiVersion v2.0.0
     *  @apiParamExample {json} [example]
     *  {
+    *
     *  }
-    *  @apiSuccess {json} Data 返回修改后的数据
+    *  @apiSuccess {json} Data 返回app的所有数据
     *  @apiSuccessExample Success-Response:
     *    HTTP/1.1 200 OK
     *       [{
-    *        "id": "1",
+    *        "id": 1,
     *        "app_id": "1",
     *        "key": "key1"
     *        "value": "value1",
@@ -28,7 +29,7 @@ class DataController extends Controller {
     *        "created_at": "2017-08-21 16:00",
     *       },
     *       {
-    *        "id": "2",
+    *        "id": 2,
     *        "app_id": "1",
     *        "key": "key2"
     *        "value": "value2",
@@ -37,20 +38,29 @@ class DataController extends Controller {
     *       }]
     */
     public function show(Request $request, $app_id) {
-        // $this->validate($request, [
-        //     'app_id' => 'required|integer|min:1',
-        // ]);
         $app = App::withTrashed()->where('id', '=', $app_id)->first();
-        $datas = $app->data()->get();
-        return $datas;
+        if(! $app) abort(404);
+        $datas = $app->data;
+        $response = [];
+        foreach($datas as $key => $data) {
+            $response[$key] = [
+                'id' => $data->id,
+                'app_id' => $app_id,
+                'key' => $data->key,
+                'value' => $data->value,
+                'created_at' => $data->created_at->timestamp,
+                'updated_at' => $data->updated_at->timestamp,
+            ];
+        }
+        return $response;
     }
     /**
-    *  @api {post} /api/{app_id}/data 添加app数据
+    *  @api {post} /api/app/{app_id}/data 添加app数据
     *  @apiName add_data
     *  @apiGroup Data
-    *  @apiVersion v1.0.0
-    *  @apiParam (must) {string} key 数据的关键词
-    *  @apiParam (must) {string} value 数据的值
+    *  @apiVersion v2.0.0
+    *  @apiParam (MUST) {string} key 数据的关键词
+    *  @apiParam (MUST) {string} value 数据的值
     *  @apiParamExample {json} [example]
     *  {
     *    "key": "key3",
@@ -60,7 +70,7 @@ class DataController extends Controller {
     *  @apiSuccessExample Success-Response:
     *    HTTP/1.1 200 OK
     *       {
-    *        "id": "3",
+    *        "id": 3,
     *        "app_id": "1",
     *        "key": "key3"
     *        "value": "value3",
@@ -70,24 +80,34 @@ class DataController extends Controller {
     */
     public function store(Request $request, $app_id) {
         $this->validate($request, [
-            // 'app_id' => 'required|integer|min:1',
             'key' => 'required|string',
             'value' => 'required|string',
         ]);
         $data = new Data;
+        $has = Data::where('app_id', '=', $app_id)->where('key', '=', $request->key)->first();
+        if($has !== null) abort(555);
         $data->app_id = $app_id;
         $data->key = $request->key;
         $data->value = $request->value;
         $data->save();
-        return $data;
+        $response = [
+            'id' => $data->id,
+            'app_id' => $app_id,
+            'key' => $data->key,
+            'value' => $data->value,
+            'created_at' => $data->created_at->timestamp,
+            'updated_at' => $data->updated_at->timestamp,
+        ];
+        return $response;
     }
     /**
-    *  @api {put} /api/{app_id}/data 修改app数据
+    *  @api {put} /api/app/{app_id}/data 修改app数据
     *  @apiName change_data
     *  @apiGroup Data
-    *  @apiVersion v1.0.0
-    *  @apiParam (nullable) {string} key 数据的关键词
-    *  @apiParam (nullable) {string} value 数据的值
+    *  @apiVersion v2.0.0
+    *  @apiParam (MUST) {integer} data_id 要修改的data的id
+    *  @apiParam (Nullable) {string} key 数据的关键词
+    *  @apiParam (Nullable) {string} value 数据的值
     *  @apiParamExample {json} [example]
     *  {
     *    "key": "key3_change",
@@ -97,7 +117,7 @@ class DataController extends Controller {
     *  @apiSuccessExample Success-Response:
     *    HTTP/1.1 200 OK
     *       {
-    *        "id": "3",
+    *        "id": 3,
     *        "app_id": "1",
     *        "key": "key3_change"
     *        "value": "value3_change",
@@ -107,22 +127,33 @@ class DataController extends Controller {
     */
     public function change(Request $request, $app_id) {
         $this->validate($request, [
-            // 'app_id' => 'required|integer|min:1',
             'data_id' => 'required|integer|min:1',
             'key' => 'required|string',
             'value' => 'required|string',
         ]);
         $data = Data::find($request['data_id']);
-        $data->key = $request->key;
-        $data->value = $request->value;
+        if(isset($request['key']) && $request['key'] !== null) {
+            $has = Data::where('app_id', '=', $app_id)->where('key', '=', $request->key)->first();
+            if($has !== null && $has->id != $request->data_id) abort(555);
+            $data->key = $request->key;
+        }
+        if(isset($request['value']) && $request['value'] !== null) $data->value = $request->value;
         $data->save();
-        return $data;
+        $response = [
+            'id' => $data->id,
+            'app_id' => $app_id,
+            'key' => $data->key,
+            'value' => $data->value,
+            'created_at' => $data->created_at->timestamp,
+            'updated_at' => $data->updated_at->timestamp,
+        ];
+        return $response;
     }
     /**
-    *  @api {put} /api/{app_id}/data 删除app数据
+    *  @api {put} /api/app/{app_id}/data 删除app数据
     *  @apiName delete_data
     *  @apiGroup Data
-    *  @apiVersion v1.0.0
+    *  @apiVersion v2.0.0
     *  @apiParam (must) {integer} data_id 数据的id
     *  @apiParamExample {json} [example]
     *  {
@@ -135,10 +166,10 @@ class DataController extends Controller {
     */
     public function delete(Request $request, $app_id) {
         $this->validate($request, [
-            // 'app_id' => 'required|integer|min:1',
             'data_id' => 'required|integer|min:1',
         ]);
         $data = Data::find($request->data_id);
+        if(! $data) abort(404);
         $data->delete();
         return [];
     }
